Move omc4py.session.string -> omc4py.string
