+++ omc4py/string.py
